refactor(node): rename upgradable.ts → gen-server.ts (and its test)
The file started as hot-code-upgrades and grew into the whole gen_server
implementation (genServer, Behavior, GenServer, Self, superviseGenServer,
links). Rename it — and its test — to say what it is. Pure move: the barrel is
the only importer, and tests reach it through the barrel.
